schema/sched: Use uint64
This commit updates the sched package to use uint64 instead of integers

+++ sema/sched_test.go
@@ -13,7 +13,7 @@ func ExampleSchedule() {
 	var counter atomic.Uint64
 
 	// because we have a parallelism of 1, we run exactly in order!
-	_ = Schedule(func(i int) error {
+	_ = Schedule(func(i uint64) error {
 		counter.Add(1)
 		return nil
 	}, 1000, Concurrency{
@@ -27,7 +27,7 @@ func ExampleSchedule() {
 
 func ExampleSchedule_order() {
 	// because we have a parallelism of 1, we run exactly in order!
-	_ = Schedule(func(i int) error {
+	_ = Schedule(func(i uint64) error {
 		fmt.Print(i, ";")
 		return nil
 	}, 4, Concurrency{
@@ -39,7 +39,7 @@ func ExampleSchedule_order() {
 }
 
 func ExampleSchedule_error() {
-	err := Schedule(func(i int) error {
+	err := Schedule(func(i uint64) error {
 		// the first invocation produces an error and returns immediately!
 		if i == 0 {
 			return errors.New("first function error")
@@ -68,7 +68,7 @@ func ExampleSchedule_error() {
 func ExampleSchedule_force() {
 	var counter atomic.Uint64
 
-	err := Schedule(func(i int) error {
+	err := Schedule(func(i uint64) error {
 		// count the number of invocations
 		counter.Add(1)
